1) Change float to double in temperature 2) Switch off payselection buttons for pay 3) Add change temperature functionality 4) Fix MyAccount view

+++ src/main/java/ru/rusguardian/domain/user/AISettingsEmbedded.java
@@ -19,7 +19,7 @@ public class AISettingsEmbedded {
     @Enumerated(EnumType.STRING)
     @Column(name = "ai_language")
     private AILanguage aiLanguage;
-    private float temperature;
+    private double temperature;
     @Enumerated(EnumType.STRING)
     @Column(name = "ai_active_model")
     private AIModel aiActiveModel;
